fix(conform): run formatters or fall back to using LSPs and nvim-lsp
Due to a lack of understanding the interplay between parsers,
formatters, linters and the conform + nvim-list/nvim-lsp combination I
ran into issues. Specifically when trying to use eslint while making
sure conform.nvim would use Prettier if installed for this project. I
hate Prettier but some teams insist on using it so it should be
prioritized when present.

Conform runs and/or tries the configured formatters in sequence and will
fall back to nvim-lsp if unavailable.