refactor(hyprland): pull apart into separate module files
While there is a lot of boilerplate involved to setup separate module
files for everything, I'm beginning to see the value in abstracting
everything away into their own components. Even if there isn't much to
setup or install everything relevant can be bundled together and
complexity is reduced as a result of that.

This commit also introduced uwsm to set the correct environment
variables in an attempt to fix screen sharing issues. Unfortunately to
no avail.. uwsm will require me to do some more reading and configuring.